Removed duplicate models and warnings that came along with it
